Moving PageObjects to another package
