+++ src/test/java/eu/yals/test/pageobjects/HomePageObject.java
@@ -1,4 +1,4 @@
-package eu.yals.test.ui.pageobjects;
+package eu.yals.test.pageobjects;
 
 import com.codeborne.selenide.SelenideElement;
 import eu.yals.test.utils.vaadin.elements.ButtonElement;
