+++ src/test/java/eu/yals/test/pageobjects/external/Eki.java
@@ -1,4 +1,4 @@
-package eu.yals.test.ui.pageobjects.external;
+package eu.yals.test.pageobjects.external;
 
 /**
  * Page object for page http://eki.ee/dict/ekss/index.cgi
