+++ src/test/java/eu/yals/test/ui/usage/IdnTest.java
@@ -1,9 +1,9 @@
 package eu.yals.test.ui.usage;
 
 import com.codeborne.selenide.SelenideElement;
+import eu.yals.test.pageobjects.HomePageObject;
+import eu.yals.test.pageobjects.external.*;
 import eu.yals.test.ui.SelenideTest;
-import eu.yals.test.ui.pageobjects.HomePageObject;
-import eu.yals.test.ui.pageobjects.external.*;
 import eu.yals.test.utils.SelenideUtils;
 import org.junit.Assert;
 import org.junit.Before;
